feat: 도메인로직 ts로 마이그레이션
