+++ src/components/modal/RestaurantAddModal.js
@@ -1,11 +1,10 @@
 import createDOMElement from '../../util/createDomElement.js';
 import { $ } from '../../util/selector.js';
 import Modal from '../Modal.js';
-import RestaurantIcon from '../restaurant/RestaurantIcon.js';
-import RestaurantItem from '../restaurant/RestaurantItem.js';
+
 import RestaurantAddModalForm from './RestaurantAddModalForm.js';
 
-function RestaurantAddModal() {
+function RestaurantAddModal({ onSubmit }) {
   const modal = new Modal({
     content: createDOMElement({
       tag: 'div',
@@ -18,9 +17,8 @@ function RestaurantAddModal() {
         }),
         RestaurantAddModalForm({
           onClose: () => modal.close(),
-
           onSubmit: (e) => {
-            handleSubmit(e);
+            onSubmit(e);
             modal.close();
           }
         })
@@ -30,22 +28,5 @@ function RestaurantAddModal() {
 
   return modal;
 }
-
-const handleSubmit = (e) => {
-  e.preventDefault();
-  const formData = new FormData(e.target);
-  const { name, distance, description, category } = Object.fromEntries(formData.entries());
-
-  const item = RestaurantItem({
-    name,
-    distance,
-    description,
-    icon: RestaurantIcon({ src: `images/category-${category}.png`, alt: category })
-  });
-
-  $('.restaurant-list').appendChild(item);
-
-  e.target.reset();
-};
 
 export default RestaurantAddModal;
